Update enets2.php
This class will assist you to create the necessary payload and calculate hmac required. 

At the moment, it also supports creation of the HTML to execute the script required and process the result (both through redirection and backend notification) in your web integration project (this is also applicable if you are using webview for your mobile apps). 

Additional handling would be needed for server to server scenario, an also when Mobile SDK is involved.